+++ src/exceptions/ShellCommandException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace nstcactus\craftcms\diskUsageWidget\exceptions;
+
+use Exception;
+
+class ShellCommandException extends Exception
+{
+}