+++ src/translations/fr/disk-usage-widget.php
@@ -1,6 +1,8 @@
 <?php return [
     'Disk usage' => 'Espace disque utilisé',
     'Disk usage in the {directory} directory' => 'Espace disque utilisé dans le dossier {directory}',
+    'Disk usage in the {partition} partition' => 'Espace disque utilisé dans la partition {partition}',
+    'No partition having quota found' => 'Aucune partition ayant un quota trouvée',
     'Directory' => 'Dossier',
     'The widget will display the disk usage in the partition in which this directory resides.' => "Ce widget affiche l'espace disque utilisé sur la partition contenant ce dossier.",
     'Soft limit' => 'Limit "soft"',
@@ -8,5 +10,10 @@
     'This field must be the path to an existing directory on the server' => 'Ce champ doit être le chemin vers un dossier existant sur le serveur.',
     'Disk usage: {used} / {total}' => 'Espace disque utilisé : {used} / {total}',
     'Soft-limit: {amount}' => 'Limite "soft" : {amount}',
-    'sloubi' => 'Sloubi un, Sloubi deux…',
+    'In quotas mode, the soft limit is set by the hosting provider. Enable this to set your own limit.' => "En mode quotas, la limit \"soft\" est définie par l'hébergeur. Activez ceci pour définir votre propre limite.",
+    'The server uses disk quotas' => "Le serveur utilise des quotas d'espace disque",
+    'If the server uses disk quotas, the output of the <code>quota</code> command can be used to display more accurate information.' => "Si le serveur utilise des quotas d'espace disque, la commande <code>quota</code> peut être utilisée pour afficher des informations plus précises.",
+    'Partition' => 'Partition',
+    'Override soft limit' => "Définir une limit \"soft\" personnalisée",
+    'The widget will display the disk usage in this partition.' => "Le widget affiche l'espace disque utilisé sur cette partition.",
 ];